# frontend/Agricultural_Products_Market_admin/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/admin/user/status": {
      "put": {
        "tags": [
          "管理员用户管理"
        ],
        "summary": "禁用/启用用户账号",
        "description": "更新用户账号状态",
        "operationId": "updateUserStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/produce/{id}": {
      "put": {
        "tags": [
          "管理员农产品管理"
        ],
        "summary": "修改农产品",
        "description": "修改指定ID的农产品信息",
        "operationId": "updateProduce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduceRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "管理员农产品管理"
        ],
        "summary": "删除农产品",
        "description": "删除指定ID的农产品",
        "operationId": "deleteProduce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/admin/produce/status": {
      "put": {
        "tags": [
          "管理员农产品管理"
        ],
        "summary": "上架/下架农产品",
        "description": "更新农产品的上架状态",
        "operationId": "updateProduceStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProduceStatusDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/produce": {
      "get": {
        "tags": [
          "管理员农产品管理"
        ],
        "summary": "分页查询所有农产品",
        "description": "分页查询所有农产品，支持按名称模糊搜索",
        "operationId": "getProduceList_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "produceName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIPageProduceInfo"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "管理员农产品管理"
        ],
        "summary": "添加农产品",
        "description": "添加新的农产品信息",
        "operationId": "addProduce",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduceRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/produce": {
      "get": {
        "tags": [
          "农产品管理"
        ],
        "summary": "获取农产品列表",
        "description": "获取上架农产品列表，支持按名称模糊查询",
        "operationId": "getProduceList",
        "parameters": [
          {
            "name": "produceName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "currentPage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageDTOListProduceInfo"
                }
              }
            }
          }
        }
      }
    },
    "/produce/{id}": {
      "get": {
        "tags": [
          "农产品管理"
        ],
        "summary": "获取农产品详情",
        "description": "获取指定ID的农产品详细信息",
        "operationId": "getProduceDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProduceInfo"
                }
              }
            }
          }
        }
      }
    },
    "/admin/user": {
      "get": {
        "tags": [
          "管理员用户管理"
        ],
        "summary": "查看所有普通用户列表",
        "description": "分页查询所有普通用户",
        "operationId": "getUserList",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIPageSysUser"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
...
      "UpdateUserStatusDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProduceRequestDTO": {
        "type": "object",
        "properties": {
          "produceName": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "manufacturer": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "coverUrl": {
            "type": "string"
          }
        }
      },
      "UpdateProduceStatusDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProduceInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "农产品ID",
            "format": "int64"
          },
          "produceName": {
            "type": "string",
            "description": "农产品名称"
          },
          "categoryId": {
            "type": "integer",
            "description": "分类ID",
            "format": "int64"
          },
          "manufacturer": {
            "type": "string",
            "description": "生产商"
          },
          "price": {
            "type": "number",
            "description": "农产品价格"
          },
          "stock": {
            "type": "integer",
            "description": "库存数量",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "description": "农产品简介"
          },
          "status": {
            "type": "integer",
            "description": "状态：1-上架，0-下架",
            "format": "int32"
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图片URL"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "农产品信息表"
      },
      "PageDTOListProduceInfo": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "总条数",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "description": "总页码数",
            "format": "int64"
          },
          "list": {
            "type": "array",
            "description": "当前页数据",
            "items": {
              "type": "array",
              "description": "当前页数据",
              "items": {
                "$ref": "#/components/schemas/ProduceInfo"
              }
            }
          }
        },
        "description": "分页结果"
      },
      "ResultPageDTOListProduceInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageDTOListProduceInfo"
          }
        }
      },
      "ResultProduceInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ProduceInfo"
          }
        }
      },
      "IPageSysUser": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SysUser"
            }
          },
          "current": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ResultIPageSysUser": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/IPageSysUser"
          }
        }
      },
      "SysUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "username": {
            "type": "string",
            "description": "登录用户名"
          },
          "password": {
            "type": "string",
            "description": "加密密码（BCrypt）"
          },
          "role": {
            "type": "string",
            "description": "角色：ROLE_VISITOR/ROLE_USER/ROLE_ADMIN"
          },
          "status": {
            "type": "integer",
            "description": "账号状态：1-启用，0-禁用",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "用户信息表"
      },
...
      "IPageProduceInfo": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduceInfo"
            }
          },
          "current": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ResultIPageProduceInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/IPageProduceInfo"
          }
        }
      },
...
    }
  },
...
}